feat: Update presensi-lab-pemrograman and presensi-lab-aplikasi-komputasi views
This commit updates the presensi-lab-pemrograman.blade.php and presensi-lab-aplikasi-komputasi.blade.php views to improve the clarity of the user interface. The changes include updating the text for the presensi (attendance) process and adding a message for when a user has already attended in another room. These updates enhance the user experience and provide clearer instructions for the presensi process.